Fix for not setting Fixture.MatchStatus to string representation of MatchStatusEnum value
The enum must first be casted to int, which was missed in this case. Example:
MatchStatus.InRunning.ToString() => "InRunning", error!
((int)MatchStatus.InRunning).ToString() => "40", what we want.

Also adding an unittest that detects the error.

Ref: #refs FS-2002

+++ SS.Integration.Adapter/StreamListener.cs
@@ -1060,7 +1060,7 @@ namespace SS.Integration.Adapter
             else
             {
                 if (state != null)
-                    fixture.MatchStatus = state.MatchStatus.ToString();
+                    fixture.MatchStatus = ((int)state.MatchStatus).ToString();
 
                 try
                 {
